Tests

+++ src/main/java/br/ufpb/dcx/apps4society/educapi/services/ContextService.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 import br.ufpb.dcx.apps4society.educapi.domain.User;
 import br.ufpb.dcx.apps4society.educapi.dto.context.ContextRegisterDTO;
 import br.ufpb.dcx.apps4society.educapi.repositories.UserRepository;
-import br.ufpb.dcx.apps4society.educapi.services.exceptions.InvalidUserException;
+import br.ufpb.dcx.apps4society.educapi.services.exceptions.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,18 +17,26 @@ import org.springframework.transaction.annotation.Transactional;
 import br.ufpb.dcx.apps4society.educapi.domain.Context;
 import br.ufpb.dcx.apps4society.educapi.dto.context.ContextDTO;
 import br.ufpb.dcx.apps4society.educapi.repositories.ContextRepository;
-import br.ufpb.dcx.apps4society.educapi.services.exceptions.ObjectNotFoundException;
 
 @Service
 public class ContextService {
+    
     @Autowired
-    private JWTService jwtService;
+    //private
+    public JWTService jwtService;
 
     @Autowired
     private ContextRepository contextRepository;
 
     @Autowired
     private UserRepository userRepository;
+
+    // para testes com ServicesBuilder
+    public ContextService(JWTService jwtService, ContextRepository contextRepository, UserRepository userRepository) {
+        this.jwtService = jwtService;
+        this.contextRepository = contextRepository;
+        this.userRepository = userRepository;
+    }
 
     public Context find(Long id) throws ObjectNotFoundException {
 
@@ -40,12 +48,21 @@ public class ContextService {
     }
 
     @Transactional
-    public ContextDTO insert(String token, ContextRegisterDTO contextRegisterDTO) throws ObjectNotFoundException, InvalidUserException {
+    public ContextDTO insert(String token, ContextRegisterDTO contextRegisterDTO) throws ContextAlreadyExistsException, ObjectNotFoundException, InvalidUserException {
         User user = validateUser(token);
 
-        Context context = contextRegisterDTO.toContext();
+        Context context = contextRegisterDTO.contextRegisterDTOToContext();
 
         context.setCreator(user);
+
+        // Início de trecho a verificar necessidade
+        Optional<Context> contextOptional = contextRepository.findContextByNameIgnoreCase(context.getName());
+
+        if (contextOptional.isPresent()){
+            throw new ContextAlreadyExistsException("There is already a context with this name registered in the system!");
+        }
+        // Fim de trecho a verificar necessidade
+
         contextRepository.save(context);
         return new ContextDTO(context);
     }
@@ -53,12 +70,21 @@ public class ContextService {
     public ContextDTO update(String token, ContextRegisterDTO contextRegisterDTO, Long id) throws ObjectNotFoundException, InvalidUserException {
         User user = validateUser(token);
 
+        // Início de trecho a verificar necessidade
+        Optional<Context> contextOptional = contextRepository.findById(id);
+
+        if (!contextOptional.isPresent()){
+            throw new ObjectNotFoundException();
+        }
+        // Fim de trecho a verificar necessidade
+
         Context newObj = find(id);
         if (!newObj.getCreator().equals(user)) {
-            throw new InvalidUserException();
+            throw new InvalidUserException("User: " + user.getName() + " is not the owner of the context: "
+                    + newObj.getName() + ".");
         }
 
-        updateData(newObj, contextRegisterDTO.toContext());
+        updateData(newObj, contextRegisterDTO.contextRegisterDTOToContext());
         contextRepository.save(newObj);
         return new ContextDTO(newObj);
     }
@@ -66,14 +92,23 @@ public class ContextService {
     public ContextDTO delete(String token, Long id) throws ObjectNotFoundException, InvalidUserException {
         User user = validateUser(token);
 
+        // Início de trecho a verificar necessidade
+        Optional<Context> contextOptional = contextRepository.findById(id);
+
+        if (!contextOptional.isPresent()){
+            throw new ObjectNotFoundException();
+        }
+        // Fim de trecho a verificar necessidade
+
         Context context = find(id);
         if (!context.getCreator().equals(user)) {
-            throw new InvalidUserException();
+            throw new InvalidUserException("User: " + user.getName() + " is not the owner of the context: "
+                    + context.getName() + ".");
         }
+
         contextRepository.deleteById(id);
         return new ContextDTO(context);
     }
-
 
     public Page<Context> findContextsByParams(String email, String name, Pageable pageable) {
         if (email != null && name != null){
@@ -87,10 +122,13 @@ public class ContextService {
         }
     }
 
-    public List<ContextDTO> findContextsByCreator(String token) throws ObjectNotFoundException, InvalidUserException {
+    public List<ContextDTO> findContextsByCreator(String token) throws  ObjectNotFoundException, InvalidUserException {
         User user = validateUser(token);
 
         List<Context> contextListByCreator = contextRepository.findContextsByCreator(user);
+        if(contextListByCreator.isEmpty()) {
+            throw new ObjectNotFoundException();
+        }
 
         return contextListByCreator.stream().map(ContextDTO::new).collect(Collectors.toList());
     }
